EEGMamba/requirements.txt
einops
h5py
lmdb
matplotlib
mne
numpy
pandas
ptflops
pyEDFlib
scikit_learn
scipy
torchinfo
tqdm
wfdb
